+++ practice3/src/buildings/net/server/parallel/BinaryServer.java
@@ -18,7 +18,7 @@ public class BinaryServer {
             throw new BuildingUnderArrestException();
 
         double squarePrice;
-        int square = building.getSumSquare();
+        double square = building.getSumSquare();
         if (building instanceof Dwelling) {
             squarePrice = 1000;
         } else if (building instanceof OfficeBuilding) {
